refactor(landing): moved out of index to a feature
