+++ app/client/src/features/landing/Carousel/Carousel.tsx
